typewit::type_fn

Struct FnRev

pub struct FnRev<F: ?Sized>(/* private fields */);

Reverses an InjTypeFn, its arguments become return values, and its return values become arguments.

Examples

Permutations

The different ways this function can be combined with CallFn and UncallFn

use typewit::type_fn::{CallFn, FnRev, UncallFn};
 
let _: CallFn<FnRev<Swap>, Right> = Left;
let _: UncallFn<    Swap,  Right> = Left;
 
let _: CallFn<        Swap,  Up> = Down;
let _: UncallFn<FnRev<Swap>, Up> = Down;
 
typewit::inj_type_fn!{
    struct Swap;
 
    impl Left => Right;
    impl Up   => Down;
}
 
struct Left;
struct Right;
struct Up;
struct Down;

Implementations

impl<F: ?Sized> FnRev<F>

pub const NEW: Self

Constructs a FnRev.

impl<F> FnRev<F>

pub const fn from_ref(_f: &F) -> Self

Constructs a FnRev from &F